[Bugfix] Resolve operator name collision for DeepSeekV3.2 in RL scenarios
