+++ csrc/lightning_indexer_custom/op_host/CMakeLists.txt
@@ -8,7 +8,7 @@
 # ======================================================================================================================
 
 add_ops_compile_options(
-        OP_NAME LightningIndexer
+        OP_NAME LightningIndexerCustom
         OPTIONS --cce-auto-sync=off
                 -Wno-deprecated-declarations
                 -Werror
@@ -16,19 +16,19 @@ add_ops_compile_options(
                 --op_relocatable_kernel_binary=true
 )
 
-set(lightning_indexer_depends transformer/attention/lightning_indexer PARENT_SCOPE)
+set(lightning_indexer_custom_depends transformer/attention/lightning_indexer_custom PARENT_SCOPE)
 
 target_sources(op_host_aclnn PRIVATE
-        lightning_indexer_def.cpp
+        lightning_indexer_custom_def.cpp
 )
 
 target_sources(optiling PRIVATE
-        lightning_indexer_tiling.cpp
+        lightning_indexer_custom_tiling.cpp
 )
 
 if (NOT BUILD_OPEN_PROJECT)
     target_sources(opmaster_ct PRIVATE
-        lightning_indexer_tiling.cpp
+        lightning_indexer_custom_tiling.cpp
     )
 endif ()
 
@@ -37,6 +37,6 @@ target_include_directories(optiling PRIVATE
 )
 
 target_sources(opsproto PRIVATE
-        lightning_indexer_proto.cpp
+        lightning_indexer_custom_proto.cpp
 )
