+++ csrc/torch_binding.cpp
@@ -683,7 +683,7 @@ at::Tensor& dispatch_ffn_combine(
     return out;
 }
 
-at::Tensor npu_lightning_indexer(
+at::Tensor npu_lightning_indexer_custom(
     const at::Tensor &query, const at::Tensor &key, const at::Tensor &weights,
     const c10::optional<at::Tensor> &actual_seq_lengths_query,
     const c10::optional<at::Tensor> &actual_seq_lengths_key,
@@ -716,12 +716,12 @@ at::Tensor npu_lightning_indexer(
         n_dim_index = (key_layout_str == "TND") ? DIM_1 : DIM_2;
         output_size = {query.size(DIM_0), key.size(n_dim_index), sparse_count};
     }
-    at::Tensor lightning_indexer_output = at::empty(output_size, query.options().dtype(at::kInt));
+    at::Tensor lightning_indexer_custom_output = at::empty(output_size, query.options().dtype(at::kInt));
     // convert str
     char *query_layout_ptr = const_cast<char *>(query_layout_str.c_str());
     char *key_layout_ptr = const_cast<char *>(key_layout_str.c_str());
     EXEC_NPU_CMD(
-        aclnnLightningIndexer,
+        aclnnLightningIndexerCustom,
         query,
         key,
         weights,
@@ -732,11 +732,11 @@ at::Tensor npu_lightning_indexer(
         key_layout_ptr,
         sparse_count,
         sparse_mode,
-        lightning_indexer_output);
-    return lightning_indexer_output;
+        lightning_indexer_custom_output);
+    return lightning_indexer_custom_output;
 }
 
-at::Tensor npu_sparse_flash_attention(
+at::Tensor npu_sparse_flash_attention_custom(
     const at::Tensor &query, const at::Tensor &key, const at::Tensor &value,
     const at::Tensor &sparse_indices, double scale_value, int64_t sparse_block_size,
     const c10::optional<at::Tensor> &block_table,
@@ -761,7 +761,7 @@ at::Tensor npu_sparse_flash_attention(
     char *layout_kv_ptr = const_cast<char *>(layout_kv_str.c_str());
 
     EXEC_NPU_CMD(
-        aclnnSparseFlashAttention,
+        aclnnSparseFlashAttentionCustom,
         query,
         key,
         value,
@@ -1320,22 +1320,22 @@ TORCH_LIBRARY_EXPAND(CONCAT(_C, _ascend), ops)
     ops.impl("grouped_matmul_swiglu_quant_weight_nz_tensor_list", torch::kPrivateUse1, &vllm_ascend::grouped_matmul_swiglu_quant_weight_nz_tensor_list);
 
     ops.def(
-        "npu_lightning_indexer(Tensor query, Tensor key, Tensor weights, *,"
+        "npu_lightning_indexer_custom(Tensor query, Tensor key, Tensor weights, *,"
         "                      Tensor? actual_seq_lengths_query=None, Tensor? actual_seq_lengths_key=None,"
         "                      Tensor? block_table=None, str layout_query='BSND', str layout_key='BSND',"
         "                      int sparse_count=2048, int sparse_mode=3) -> Tensor"
     );
-    ops.impl("npu_lightning_indexer", torch::kPrivateUse1, &vllm_ascend::npu_lightning_indexer);
+    ops.impl("npu_lightning_indexer_custom", torch::kPrivateUse1, &vllm_ascend::npu_lightning_indexer_custom);
 
     ops.def(
-        "npu_sparse_flash_attention(Tensor query, Tensor key, Tensor value,"
+        "npu_sparse_flash_attention_custom(Tensor query, Tensor key, Tensor value,"
         "                           Tensor sparse_indices, float scale_value, int sparse_block_size, *,"
         "                           Tensor? block_table=None, Tensor? actual_seq_lengths_query=None,"
         "                           Tensor? actual_seq_lengths_kv=None, Tensor? query_rope=None,"
         "                           Tensor? key_rope=None, str layout_query='BSND', str layout_kv='BSND',"
         "                           int sparse_mode=3) -> Tensor"
     );
-    ops.impl("npu_sparse_flash_attention", torch::kPrivateUse1, &vllm_ascend::npu_sparse_flash_attention);
+    ops.impl("npu_sparse_flash_attention_custom", torch::kPrivateUse1, &vllm_ascend::npu_sparse_flash_attention_custom);
 
     ops.def(
         "dispatch_ffn_combine(Tensor x, Tensor[] weight1, Tensor[] weight2, Tensor expert_idx,"
